refactored `reload` - moved methods into separate file
